requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.31.0
streamlit>=1.25.0
Pillow>=9.0.0
numpy>=1.24.0
accelerate>=0.21.0
